# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Packaging Compliance AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fda/recalls:
    get:
      operationId: listFdaRecalls
      tags: [regulations]
      summary: Search live FDA recall & enforcement actions (openFDA)
      parameters:
        - name: category
          in: query
          required: true
          schema:
            type: string
            enum: [food, drug, device]
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Matching FDA recalls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FdaRecallResponse"
...
  /fda/intelligence:
    get:
      operationId: getFdaIntelligence
      tags: [regulations]
      summary: Assemble applicable FDA regulatory intelligence for a package
      parameters:
        - name: packageId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: FDA intelligence for the package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FdaIntelligence"
...
  /ai-providers:
    get:
      operationId: listAiProviders
      tags: [ai]
      summary: List configured AI model providers
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiProvider"
    post:
      operationId: createAiProvider
      tags: [ai]
      summary: Add an AI model provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiProviderInput"
      responses:
        "201":
          description: Created provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProvider"
  /ai-providers/{id}:
    patch:
      operationId: updateAiProvider
      tags: [ai]
      summary: Update an AI model provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiProviderUpdate"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProvider"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteAiProvider
      tags: [ai]
      summary: Remove an AI model provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /ai-providers/{id}/activate:
    post:
      operationId: activateAiProvider
      tags: [ai]
      summary: Make this provider the active analysis engine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProvider"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /ai-providers/{id}/test:
    post:
      operationId: testAiProvider
      tags: [ai]
      summary: Test connectivity to an AI model provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProviderTestResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    FdaRecall:
      type: object
      required:
        [
          recallNumber,
          category,
          status,
          classification,
          productDescription,
          reason,
          recallingFirm,
          distributionPattern,
          location,
        ]
      properties:
        recallNumber:
          type: string
        category:
          type: string
        status:
          type: string
        classification:
          type: string
        productDescription:
          type: string
        reason:
          type: string
        recallingFirm:
          type: string
        distributionPattern:
          type: string
        location:
          type: string
        initiationDate:
          type: string
          nullable: true
        reportDate:
          type: string
          nullable: true
    FdaRecallResponse:
      type: object
      required: [results, total, disclaimer]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/FdaRecall"
        total:
          type: integer
        disclaimer:
          type: string
    FdaSource:
      type: object
      required: [id, label, description]
      properties:
        id:
          type: string
        label:
          type: string
        description:
          type: string
    FdaFinding:
      type: object
      required: [source, title, detail]
      properties:
        source:
          type: string
        title:
          type: string
        detail:
          type: string
        severity:
          type: string
        date:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
    FdaLabelExample:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
        url:
          type: string
          nullable: true
    FdaReference:
      type: object
      required: [code, title]
      properties:
        code:
          type: string
        title:
          type: string
    FdaLink:
      type: object
      required: [label, url]
      properties:
        label:
          type: string
        url:
          type: string
...
    FdaIntelligence:
      type: object
      required:
        [
          detectedCategory,
          categoryLabel,
          searchTerm,
          applicableSources,
          warnings,
          findings,
          labelExamples,
          references,
          sourceLinks,
          available,
          degraded,
          message,
          disclaimer,
        ]
      properties:
        detectedCategory:
          type: string
          nullable: true
        categoryLabel:
          type: string
        searchTerm:
          type: string
          nullable: true
        applicableSources:
          type: array
          items:
            $ref: "#/components/schemas/FdaSource"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/FdaFinding"
        findings:
          type: array
          items:
            $ref: "#/components/schemas/FdaFinding"
        labelExamples:
          type: array
          items:
            $ref: "#/components/schemas/FdaLabelExample"
        references:
          type: array
          items:
            $ref: "#/components/schemas/FdaReference"
        sourceLinks:
          type: array
          items:
            $ref: "#/components/schemas/FdaLink"
        available:
          type: boolean
        degraded:
          type: boolean
        message:
          type: string
          nullable: true
        disclaimer:
          type: string
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AiProvider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        providerType:
          type: string
        model:
          type: string
        baseUrl:
          type: ["string", "null"]
        managed:
          type: boolean
        active:
          type: boolean
        status:
          type: string
        statusMessage:
          type: ["string", "null"]
        hasKey:
          type: boolean
        keyLast4:
          type: ["string", "null"]
        lastTestedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - providerType
        - model
        - managed
        - active
        - status
        - hasKey
        - createdAt
    AiProviderInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        providerType:
          type: string
        model:
          type: string
          minLength: 1
        baseUrl:
          type: string
        apiKey:
          type: string
      required:
        - name
        - model
    AiProviderUpdate:
      type: object
      properties:
        name:
          type: string
        model:
          type: string
        baseUrl:
          type: string
        apiKey:
          type: string
        active:
          type: boolean
    AiProviderTestResult:
      type: object
      properties:
        status:
          type: string
        message:
          type: ["string", "null"]
      required:
        - status
...